make choosing transactions to mark as committed a client responsibility
