+++ src/store/sync.rs
@@ -4,13 +4,12 @@ use crypto::{
 };
 
 use objects::{
-    accounts::AccountId,
     notes::{NoteId, NoteInclusionProof},
     BlockHeader, Digest,
 };
 use rusqlite::params;
 
-use crate::{errors::StoreError, store::transactions::TransactionFilter};
+use crate::errors::StoreError;
 
 use super::Store;
 
@@ -76,14 +75,12 @@ impl Store {
     pub fn apply_state_sync(
         &mut self,
         block_header: BlockHeader,
-        account_hash_changes: Vec<(AccountId, Digest)>,
         nullifiers: Vec<Digest>,
         committed_notes: Vec<(NoteId, NoteInclusionProof)>,
         new_mmr_peaks: MmrPeaks,
         new_authentication_nodes: &[(InOrderIndex, Digest)],
+        transactions_to_commit: &[Digest],
     ) -> Result<(), StoreError> {
-        let uncommitted_transactions = self.get_transactions(TransactionFilter::Uncomitted)?;
-
         let tx = self.db.transaction()?;
 
         // Update state sync block number
@@ -131,15 +128,10 @@ impl Store {
             )?;
         }
 
-        let note_ids: Vec<NoteId> = committed_notes.iter().map(|(id, _)| (*id)).collect();
-
         Store::mark_transactions_as_committed(
-            &uncommitted_transactions,
-            &note_ids,
-            &nullifiers,
-            &account_hash_changes,
-            block_header.block_num(),
             &tx,
+            block_header.block_num(),
+            transactions_to_commit,
         )?;
 
         // Commit the updates
